requirements.txt
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tiktoken>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pyyaml>=6.0
loguru>=0.7.0
tqdm>=4.66.0
streamlit>=1.28.0